+++ R/elicitSurvivalExtrapolation.R
@@ -197,13 +197,21 @@ elicitSurvivalExtrapolation<- function(){
                 
                 
                 tabPanel("Extrapolation plot",
+                         sidebarLayout(
+                           sidebarPanel(
+                             wellPanel(
+                               numericInput("alpha", "Credible Interval Size",
+                                            0.95, min = 0, max = 1),
+                            
+                             )),
+                           mainPanel(
                          plotOutput("extrapolationPlot")
+                           )
+                         )
+                )
                          
-                ),
-                # tabPanel("Joint distribution",
-                #          uiOutput("bivariate")
-                #          
-                # ),
+                ,
+                
                 tabPanel("Report",
                          wellPanel(
                            h5("Report content"),
@@ -502,7 +510,7 @@ elicitSurvivalExtrapolation<- function(){
     # Individual judgements tab ----
     
     initialVals1 <- reactive({
-      initialdf <- matrix(0.1*(1:5), nrow = 5, ncol = input$nExperts
+      initialdf <- matrix(c(0, 0.25, 0.5, 0.75, 1), nrow = 5, ncol = input$nExperts
                          )
       if(input$elicMethod == "quartiles"){
       rownames(initialdf) <- c("L", "Q1", "M", "Q3", "U")
@@ -679,7 +687,7 @@ elicitSurvivalExtrapolation<- function(){
       conditionalPanel(
         condition = "output.RIOtgpNumber == 1",
         textInput("values1", label = h5("QoI values"), 
-                  value = "0.2, 0.3, 0.4")
+                  value = "0.25, 0.5, 0.75")
       )
     })
     
@@ -714,7 +722,7 @@ elicitSurvivalExtrapolation<- function(){
       conditionalPanel(
         condition = "output.RIOtgpNumber == 2",
         textInput("values2", label = h5("QoI values"), 
-                  value = "0.5, 0.6, 0.7")
+                  value = "0.25, 0.5, 0.75")
       )
     })
     
@@ -1031,71 +1039,45 @@ elicitSurvivalExtrapolation<- function(){
     
     output$extrapolationPlot <- renderPlot({
       req(survivalDF(),
-          input$targetTime, myfit1() )
-      myplot <- KMextrapolate(tLower = 0,
-                              tUpper = input$truncationTime,
-                              yl =0 ,
-                              yu = 1,
-                              showExp = FALSE,
-                              tTarget = input$targetTime,
-                              breakTime = input$targetTime/4,
-                              survDf = survivalDF(),
-                              groups = levels(survivalDF()$treatment),
-                              xl = paste0("Time (", input$timeUnit, ")"),
-                              includeEbar = FALSE,
-                              includeExpRibbon = TRUE,
-                              KMCI = TRUE,
-                              fontsize = input$fs,
-                              useWeights = caseWeight_value())
+          input$targetTime, myfit1(), input$truncationTime,
+          input$alpha > 0, input$alpha < 1)
       
       if(input$RIOdist1 == "best"){
-        fqDist1 <- myfit1()$best.fitting[1, 1]
+        fqRIO1 <- myfit1()$best.fitting[1, 1]
       }else{
-        fqDist1 <- input$RIOdist1
+        fqRIO1 <- input$RIOdist1
       }
-      
-      fq1<- feedback(myfit1(),
-                     quantiles=c(0.025, 0.5, 0.975))$fitted.quantiles[, fqDist1]
-     
-      if(length(levels(survivalDF()$treatment)) == 1){
-        jitter <- 1
-      }else{
-        jitter <- 1.005
-      }
-      
-      
-      
-      p1 <- myplot + annotate("errorbar", x = jitter*input$targetTime,
-                             y = fq1[2],
-                             ymin = fq1[1],
-                             ymax = fq1[3],
-                             colour = "#F8766D",
-                             linewidth =2,
-                             width = 0.02*input$targetTime) 
       
       if(length(levels(survivalDF()$treatment)) > 1){
-        
+        group2RIO <- myfit2()
         if(input$RIOdist2 == "best"){
-          fqDist2 <- myfit2()$best.fitting[1, 1]
+          fqRIO2 <- myfit2()$best.fitting[1, 1]
         }else{
-          fqDist2 <- input$RIOdist2
+          fqRIO2 <- input$RIOdist2
         }
         
-        fq2<- feedback(myfit2(),
-                       quantiles=c(0.025,0.5,
-                                   0.975))$fitted.quantiles[, fqDist2]
-        p1 <- p1 +
-        annotate("errorbar", x = 0.995*input$targetTime,
-                 y = fq2[2],
-                 ymin = fq2[1],
-                 ymax = fq2[3],
-                 colour = "#00BFC4",
-                 linewidth =2,
-                 width = 0.02*input$targetTime)
+      }else{
+        group2RIO <- NULL
+        fqRIO2 <- NULL
       }
-        
-      print(p1)
+      
 
+      survivalExtrapolatePlot(survivalDF(),
+                                          myfit1 = myfit1(),
+                                          myfit2 = group2RIO,
+                                          fqDist1 = fqRIO1,
+                                          fqDist2 = fqRIO2,
+                                          tTruncate = input$truncationTime,
+                                          tTarget = input$targetTime,
+                                          alpha = input$alpha,
+                                          useWeights = caseWeight_value(),
+                                          groups = levels(survivalDF()$treatment),
+                                          xl = paste0("Time t (", input$timeUnit, ")"),
+                                          fontsize = input$fs,
+                                          breakTime = input$targetTime/8,
+                                          showPlot = TRUE,
+                                          returnPlot = FALSE)
+      
     })
 
    
@@ -1107,112 +1089,7 @@ elicitSurvivalExtrapolation<- function(){
     }) 
     
     
-    
-    # Bivariate tab ----
-    
-    m1 <- reactive({
-      req(p1(), v1())
-      approx(p1(), v1(), 0.5)$y
-    })
-    
-    m2 <- reactive({
-      req(p2(), v2())
-      approx(p2(), v2(), 0.5)$y
-    })
-    
-    df1 <- reactive({
-      req(myfit1(), myfit2(), input$concProb > 0, input$concProb < 1)
-      conc.probs <- matrix(0, 2, 2)
-      conc.probs[1, 2] <- input$concProb
-      data.frame(copulaSample(myfit1(), myfit2(), cp = conc.probs, 
-                              n = input$sampleSize, 
-                              d = c(input$RIOdist1, input$RIOdist2)))
-    })
-    
-    output$bivariate <- renderUI({
-      req(survivalDF())
-      if(length(levels(survivalDF()$treatment)) == 1){
-        h5("Single treatment group only: no joint distribution to elicit")
-      }else{
-        sidebarLayout(
-          sidebarPanel(
-            wellPanel(
-              numericInput("concProb", h5("Concordance probability"),
-                           value = 0.5,
-                           min = 0, max = 1),
-              numericInput("sampleSize", h5("Sample size"),
-                           value = 1000,
-                           min = 1),
-              downloadButton("downloadData", "Download sample")
-            )
-            
-          ),
-          mainPanel(
-            helpText("Specify the concordance probability: the probability that 
-                     the two uncertain quantities are either both below their medians,
-                     or both above their medians. The plot below is a sample from 
-                     the joint distribution, using the two RIO fitted distributions."),
-            plotOutput("bivariatePlot")
-            
-          )
-        )
-          
-      }
-    })
-   
-    
-    output$bivariatePlot <- renderPlot({
-      req(df1(), m1(), m2(), survivalDF(),
-          input$concProb > 0, input$concProb < 1 )
-      
-      
-      theme_set(theme_grey(base_size = input$fs))
-      
-      annotations <- data.frame(
-        xpos = c(Inf,Inf,-Inf,-Inf),
-        ypos =  c(Inf, -Inf,-Inf,Inf),
-        annotateText = as.character(c(input$concProb / 2, 
-                                      0.5 - input$concProb /2,
-                                      input$concProb / 2, 
-                                      0.5 - input$concProb /2)),
-        hjustvar = c(1.5, 1.5, -0.5, -0.5) ,
-        vjustvar = c(1.5, -0.5, -0.5, 1.5))
-      
-      
-      p1 <- ggplot(data = df1(), aes(x = X1, y = X2))
-    
-        p1 <- p1 + geom_point(alpha=0.15, colour = "red")
-      
-      p1 <- p1 + 
-        geom_hline(yintercept = m2())+
-        geom_vline(xintercept = m1())+
-        labs(x=levels(survivalDF()$treatment)[1],
-             y = levels(survivalDF()$treatment)[2])+
-        geom_text(data = annotations, aes(x = xpos,
-                                          y = ypos,
-                                          hjust = hjustvar,
-                                          vjust = vjustvar,
-                                          label = annotateText),
-                  size = input$fs / 2) +
-        xlim(0.95*xaxis1()[1], 1.05*xaxis1()[2])+
-        ylim(0.95*xaxis2()[1], 1.05*xaxis2()[2])
-      
-   
-        
-        suppressWarnings(suppressMessages(ggExtra::ggMarginal(p1, type = "histogram",
-                                                              fill = "red")))
-      
-      
-      
-    })
-    
-    output$downloadData <- downloadHandler(
-      filename = "joint-sample.csv",
-      content = function(file) {
-        utils::write.csv(df1(), file, row.names = FALSE)
-      }
-    )
-    
+  
     # Report writing ----
     
     output$reportGroup1 <- renderUI({
@@ -1277,7 +1154,8 @@ elicitSurvivalExtrapolation<- function(){
                        reportScenarioTest = input$reportScenarioTest,
                        expRange = expRange(),
                        reportExtrapolation = input$reportExtrapolation,
-                       useWeights = caseWeight_value())
+                       useWeights = caseWeight_value(),
+                       alpha = input$alpha)
         # Knit the document, passing in the `params` list, and eval it in a
         # child of the global environment (this isolates the code in the document
         # from the code in this app).
@@ -1362,233 +1240,9 @@ elicitSurvivalExtrapolation<- function(){
 
 # Helper function KMextrapolate ----
 
-KMextrapolate <- function(tLower = 0,
-                          tUpper,
-                          yl,
-                          yu,
-                          showExp = FALSE,
-                          expLower,
-                          expUpper,
-                          expGroup,
-                          tTarget,
-                          survDf,
-                          breakTime = NULL,
-                          groups,
-                          xl,
-                          includeEbar = TRUE,
-                          includeExpRibbon = TRUE,
-                          KMCI = FALSE,
-                          fontsize = 16,
-                          useWeights = FALSE){
-  
-  x <- ymin <- ymax <- NULL # hack to avoid R CMD check NOTE
-  
-  # Truncate data frame for plotting
-  index <- survDf$time > tUpper
-  truncatedDf <- survDf
-  truncatedDf$event[index] <- 0
-  truncatedDf$time[index] <- tUpper
-  
-  # Prepare KM plot
-  if(useWeights == TRUE){
-    fit <- survival::survfit(survival::Surv(time, event) ~ treatment,
-                             weights = weights, data = truncatedDf)}else{
-    fit <- survival::survfit(survival::Surv(time, event) ~ treatment,
-                             data = truncatedDf)  
-                             }
-  myplot <- survminer::ggsurvplot(fit, data = truncatedDf, censor = FALSE,
-                                  break.time.by = breakTime,
-                                  legend = "right",
-                                  legend.title = "",
-                                  legend.labs = groups,
-                                  xlim = c(tLower, tTarget),
-                                  xlab = xl,
-                                  conf.int = KMCI,
-                                  break.y.by = 0.2)
-  myplot$plot <- myplot$plot + geom_vline(xintercept = tTarget, linetype="dotted") +theme_bw(base_size = fontsize)
-  if(includeEbar){
-    myplot$plot <- myplot$plot + annotate("errorbar", x = 8,
-                                          y = mean(c(yu, yl)),
-                                          ymin = yu,
-                                          ymax = yl,
-                                          colour = "black")
-  }
-  
-  if(showExp){
-    
-    
-    
-    dfExp <- truncatedDf[truncatedDf$treatment == expGroup, ]
-    
-    expUpper <- min(expUpper, max(dfExp$time))
-    
-    if(useWeights == TRUE){
-      fitExp <- survival::survfit(survival::Surv(time, event) ~ 1,
-                                  weights = weights, data = dfExp)}else{
-      fitExp <- survival::survfit(survival::Surv(time, event) ~ 1, data = dfExp) 
-                               }
-    
-    
-    Plower <- summary(fitExp, times = expLower)$surv
-    index <- dfExp$time > expLower 
-    eventTruncated <- dfExp$event
-    eventTruncated[dfExp$time > expUpper] <- 0
-    eventTruncated <- eventTruncated[index]
-    timeTruncated <- dfExp$time[index]
-    timeTruncated[timeTruncated > expUpper] <- expUpper
-    
-    
-    if(useWeights == TRUE){
-      expFit <- survival::survreg(survival::Surv(time = timeTruncated-expLower,
-                                                 event = eventTruncated) ~ 1,
-                                  weights = dfExp$weights[index],
-                                  dist = "exponential", 
-                                  robust = TRUE)}else{
-      expFit <- survival::survreg(survival::Surv(time = timeTruncated-expLower,
-                                                 event = eventTruncated) ~ 1,
-                                  dist = "exponential")
-                                  }
-    
-    
-    
-    
-    lambda <- exp(-expFit$coefficients)
-    
-    myfun <- function(x) exp(-lambda*(x-expLower))*Plower
-    
-    
-    
-    myplot$plot <- myplot$plot + 
-      stat_function(fun = myfun, xlim = c(expLower, expUpper), lwd = 1.5, alpha = 0.9) +
-      stat_function(fun = myfun, xlim = c(expUpper, tTarget), linetype = "dashed")
-    
-    tVals <- seq(from = expLower, to = tTarget, length = 100)
-    mP <- summary(fitExp, times = expLower)$surv
-    sP <- summary(fitExp, times = expLower)$std.err
-    
-    mLogLambda <- summary(expFit)$table[1]
-    sLogLambda <- summary(expFit)$table[2]
-    
-    expMatrix <- matrix(0, 10000, 100)
-    
-    for(i in 1:10000){
-      randomP <- rnorm(1, mP, sP)
-      randomLambda <- exp(-rnorm(1, mLogLambda, sLogLambda ))
-      expMatrix[i, ] <- exp(-randomLambda*(tVals-expLower))*randomP
-    }
-    
-    if(includeExpRibbon){
-      
-      
-      ribbon_data <- data.frame(x =tVals,
-                                ymax = apply(expMatrix, 2, quantile, probs = 0.975),
-                                ymin = apply(expMatrix, 2, quantile, probs = 0.025),
-                                surv =tVals)
-      
-      myplot$plot <- myplot$plot +
-        geom_ribbon(data = ribbon_data,
-                    aes(x = x, ymin = ymin, ymax = ymax),
-                    fill = "gray", alpha = 0.5)
-      
-    }
-   
-    
-    
-  }
-  
-  
-  myplot$plot
-}
 
 # Helper function survivalTable ----
 
-# scenarioTestInterval <- function(tLower = 0,
-#                           tUpper,
-#                         
-#                           expLower,
-#                           expUpper,
-#                           expGroup,
-#                           tTarget,
-#                           survDf){
-#   # Truncate data frame for plotting
-#   index <- survDf$time > tUpper
-#   truncatedDf <- survDf
-#   truncatedDf$event[index] <- 0
-#   truncatedDf$time[index] <- tUpper
-#   
-#   # Prepare KM plot
-# 
-#     dfExp <- truncatedDf[truncatedDf$treatment == expGroup, ]
-#     fitExp <- survival::survfit(survival::Surv(time, event) ~ 1, data = dfExp)
-#     Plower <- summary(fitExp, times = expLower)$surv
-#     index <- dfExp$time > expLower 
-#     eventTruncated <- dfExp$event
-#     eventTruncated[dfExp$time > expUpper] <- 0
-#     eventTruncated <- eventTruncated[index]
-#     timeTruncated <- dfExp$time[index]
-#     timeTruncated[timeTruncated > expUpper] <- expUpper
-#     
-#     expFit <- survival::survreg(survival::Surv(time = timeTruncated-expLower, event = eventTruncated) ~ 1, dist = "exponential")
-#     lambda <- exp(-expFit$coefficients)
-#     
-#     mP <- summary(fitExp, times = expLower)$surv
-#     sP <- summary(fitExp, times = expLower)$std.err
-#     
-#     mLogLambda <- summary(expFit)$table[1]
-#     sLogLambda <- summary(expFit)$table[2]
-#     
-#     
-#       randomP <- rnorm(10000, mP, sP)
-#       randomLambda <- exp(-rnorm(10000, mLogLambda, sLogLambda ))
-#       tSample <- exp(-randomLambda*(tTarget-expLower))*randomP
-#    
-#     quantile(tSample, c(0.025, 0.975))
-#    
-#       
-# }
-
-
-# makeSurvivalTable <- function(survDF, breakTime, truncationTime, timeUnit, dp = 2){
-#   sv <- survival::survfit(survival::Surv(time, event) ~ treatment, data = survDF)
-#   truncationTime <- min(truncationTime, min(tapply(survDF$time, survDF$treatment, max)))
-#   sTimes <- seq(from = breakTime, to = truncationTime, by = breakTime)
-#   nTimes <- length(sTimes)
-#   nTreatments <- length(levels(survDF$treatment))
-#   tNames <- levels(survDF$treatment)
-#   
-#   pt <- matrix(round(summary(sv, times = sTimes)$surv , dp),
-#                nrow = nTimes, ncol = nTreatments)
-#   wt  <- pt
-#   wt[1, ] <- 1 - wt[1, ]
-#   if(nrow(wt) > 1){
-#     wt[-1, ] <- 1 - round(pt[2:nTimes, ]/pt[1:(nTimes - 1), ], dp)
-#     }
-#   
-#   ciLower <- matrix(round(summary(sv, times = sTimes)$lower , dp),
-#                     nrow = nTimes, ncol = nTreatments)
-#   ciUpper <- matrix(round(summary(sv, times = sTimes)$upper , dp),
-#                     nrow = nTimes, ncol = nTreatments)
-#   ci95 <-  matrix (paste0("(", ciLower, ", ", ciUpper,")"),
-#                    nrow = nTimes, ncol = nTreatments)
-#   sTable <- data.frame(paste0("[", c(0, sTimes[1:(nTimes -1)]), ",", sTimes, ")"),
-#                        pt[, 1], ci95[, 1],  wt[, 1])
-#   colnames(sTable) <- c(paste0("time interval (", timeUnit,")"),
-#                         paste0("survivor (", tNames[1], ")"),
-#                         paste0("survivor 95% CI (", tNames[1], ")"),
-#                         paste0("hazard (", tNames[1], ")"))
-#   
-#   if(nTreatments > 1){
-#     for(i in 2:nTreatments){
-#       dfTemp <-  data.frame(pt[, i], ci95[, i],  wt[, i])
-#       colnames(dfTemp) <- c(paste0("survivor (", tNames[i], ")"),
-#                             paste0("survivor 95% CI (", tNames[i], ")"),
-#                             paste0("hazard (", tNames[i], ")"))
-#       sTable <- cbind(sTable, dfTemp)
-#     }
-#   }
-#   
-#   sTable
-# }
 
 makeSurvSummaryTable <- function(survDF, sf = 3, useWeights = FALSE){