+++ R/makeSurvivalTable.R
@@ -2,15 +2,21 @@
 #' Tabulate Summary Data for Survival Extrapolation
 #'
 #' Tabulates the Kaplan Meier survivor function and within interval hazard at discrete equally spaced time points t_1,...,t_n
-#' "Within interval hazard" is defined as (1-S(t_{n+1})) / S_(t_n), using the Kaplan Meier estimate of S().
+#' "Within interval hazard" is defined as (1-S(t_[n+1])) / S_(t_n), using the Kaplan Meier estimate of S().
 #' The table is intended to be included on a summary sheet provided to experts when eliciting judgements about
 #' extrapolated survival probabilities.
 #' 
-#' @param survDf 
-#' @param breakTime 
-#' @param truncationTime 
-#' @param timeUnit 
-#' @param dp 
+#' @param survDf data frame with individual patient data. Needs three columns with names
+#' "time", "event" and "treatment" (in that order). For weighted observations (e.g. 
+#' using propensity scores), include a fourth column "weights".
+#' Values in the "event" column should be 0 for a censored observation, and 1 otherwise.
+#' The "treatment" column should be included even if there is only one treatment group.
+#' @param breakTime duration of each time interval
+#' @param truncationTime time point for the end of the last interval 
+#' @param timeUnit string variable to give unit of time
+#' @param dp number of decimal places to display
+#' @param useWeights set to TRUE if survDf includes column of weights, as described in specification 
+#' of survDf. This column is passed on to survival::survfit() as the case weights.
 #'
 #' @returns a data frame with survivor function estimates, 95% confidence intervals,
 #' and within interval hazard estimates for each time interval.
@@ -26,6 +32,8 @@
 #' makeSurvivalTable(sdf, breakTime = 50, truncationTime = 250, timeUnit = "months")
 #' }
 makeSurvivalTable <- function(survDf, breakTime, truncationTime, timeUnit, useWeights = FALSE, dp = 2){
+  
+  survDf$treatment <- as.factor(survDf$treatment)
   
   if(useWeights == TRUE){
     sv <- survival::survfit(survival::Surv(time, event) ~ treatment, weights  = weights,